pyproject.toml
[tool.poetry]
name = "universal-word"
version = "0.1"
description = "API to treat, extract and create nice web based dictionaries using LLM"
authors = ["Usama EL MOUSTAMIDE <usama@usama.pro>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
wheel = "^0.42.0"
pybind11 = "^2.11.1"
fasttext-wheel = "^0.9.2"
ebooklib = "^0.18"
nltk = "^3.8.1"
pandas = "^2.2.1"
requests = "^2.31.0"
regex = "^2023.12.25"
openai = "^1.13.3"
litestar = {extras = ["standard"], version = "^2.6.1"}
sqlalchemy = "^2.0.27"
llama-cpp-python = "^0.2.53"
python-dotenv = "^1.0.1"
devtools = "^0.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
